contributioncomment

+++ src/components/Contribution/manageContribution.js
@@ -5,10 +5,12 @@ import Table from "../utils/table/table";
 import {
   getContributionData,
   deleteContribution,
-  postContribution
+  postContribution,
+  postCommentContribution
 } from "../../actions/contribution_actions";
 import { getTermData } from "../../actions/term_actions";
 import ModalAddUpdate from "./modalAddUpdate";
+import ModalComment from "./modalComment";
 import Pagination from "../utils/pagination";
 import { getCurrentUser } from "../../services/authService";
 
@@ -18,7 +20,8 @@ class ManageContribution extends Component {
     currentPage: 1,
     searchQuery: "",
     modalShow: false,
-    modalContribution: ""
+    modalContribution: "",
+    modalComment: false
   };
 
   columns = [
@@ -34,14 +37,6 @@ class ManageContribution extends Component {
       path: "periodEdited",
       label: "Period Edited"
     },
-    // {
-    //   path: "fileURL",
-    //   label: "File URL"
-    // },
-    // {
-    //   path: "imageURL",
-    //   label: "Image URL"
-    // },
     {
       path: "comment",
       label: "Comment"
@@ -62,10 +57,23 @@ class ManageContribution extends Component {
           >
             <i className="fa fa-trash" />
           </button>
+          <button
+            onClick={() => this.handleComment(con)}
+            className="btn btn-danger btn-sm"
+          >
+            <i className="fa fa-comment" />
+          </button>
         </React.Fragment>
       )
     }
   ];
+
+  handleComment(con) {
+    this.setState({
+      modalComment: !this.state.modalComment,
+      modalContribution: con
+    })
+  }
 
   handleDisabled(con) {
     var d1 = new Date(con.periodEdited);
@@ -75,16 +83,27 @@ class ManageContribution extends Component {
   }
 
   handleDelete = con => {
-    this.props
-      .dispatch(deleteContribution(con.id))
-      .then(
-        this.props.dispatch(res =>
-          getContributionData(this.state.user.Id, this.state.user.Roles)
-        )
-      );
+    var result = window.confirm("Do you want delete this contribution?");
+    if (result) {
+      this.props
+        .dispatch(deleteContribution(con.id))
+        .then(res =>
+          this.props.dispatch(
+            getContributionData(this.state.user.Id, this.state.user.Roles)
+          )
+        );
+    }
   };
 
-  handleSubmit = (conSubmit, conId, img, word, userId, termId) => {
+  handleSubmitComment = (conId, comment, status) => {
+    this.props.dispatch(postCommentContribution(conId, comment, status))
+      .then(res =>
+        this.props.dispatch(
+          getContributionData(this.state.user.Id, this.state.user.Roles)
+        ));
+  }
+
+  handleSubmit = (conSubmit, conId, img, word, userId) => {
     var termId = this.state.term.id;
     var userId = this.state.user.Id;
     this.props
@@ -111,8 +130,10 @@ class ManageContribution extends Component {
     var d1 = new Date(this.state.term.closingDate);
     var d2 = new Date();
     d1.setDate(d1.getDate() - 15);
+    console.log(d1);
+    console.log(d2);
     this.setState({
-      disbaledAdd: d2 > d1 ? true : false
+      disbaledAdd: d2 < d1 ? true : false
     });
   };
 
@@ -126,11 +147,27 @@ class ManageContribution extends Component {
   };
 
   toggle = () => {
-    this.setState({
-      modalShow: !this.state.modalShow,
-      modalContribution: ""
-    });
+    if(this.state.disbaledAdd === false && this.state.modalContribution === ""){
+      this.setState({
+        modalShow: false,
+        modalContribution: ""
+      });
+    }
+    else{
+      this.setState({
+        modalShow: !this.state.modalShow,
+        modalContribution: ""
+      });
+    }
+   
   };
+
+  toggleComment = () => {
+    this.setState({
+      modalComment: !this.state.modalComment,
+      modalContribution: ""
+    })
+  }
 
   getData = () => {
     const { pageSize, currentPage, searchQuery } = this.state;
@@ -157,43 +194,55 @@ class ManageContribution extends Component {
     });
   };
 
+
   render() {
     const {
       pageSize,
       currentPage,
       modalContribution,
       modalShow,
-      disbaledAdd
+      disbaledAdd,
+      modalComment
     } = this.state;
     const { data: dataPagination, itemsCount: itemsCount } = this.getData();
     return (
       <div style={{ marginLeft: "50px" }}>
-        <h1 className="h3 mb-2 text-gray-800">Manage Contribution</h1>
-        <div className="row justify-content-center">
-          <div className="col-4">
+        <h1 className="h3 mb-2 text-gray-800 text-center">Manage Contribution</h1>
+        <div className="row justify-content-end">
+          <div className="col-6">
             <input
-              placeholder="Search....."
+              placeholder="Search Contribution Title"
               type="text"
               className="form-control"
               onChange={this.handleSearch}
-            />
-          </div>
+            /></div>
           <div className="col-4">
             <button
               onClick={this.toggle}
-              disabled={disbaledAdd}
+              data-toggle="collapse"
+              href="#collapseExample"
               className="btn btn-info"
             >
               Add new Contribution
             </button>
+            {disbaledAdd == false ? <div className='collapse' id='collapseExample'> <div className='card card-body'>The Term is out of date</div></div>
+              : ""}
+           
           </div>
         </div>
+        <ModalComment
+          show={modalComment}
+          toggle={this.toggleComment}
+          contributionInfo={modalContribution}
+          onSubmit={this.handleSubmitComment}
+        />
         <ModalAddUpdate
           show={modalShow}
           toggle={this.toggle}
           contributionInfo={modalContribution}
           onSubmit={this.handleSubmit}
         />
+
         <Table data={dataPagination} columns={this.columns} />
         <Pagination
           itemsCount={itemsCount}
