+++ src/components/User/manageUser.js
@@ -3,7 +3,7 @@ import { connect } from "react-redux";
 import { paginate } from '../utils/paginate';
 import Pagination from '../utils/pagination';
 import Table from "../utils/table/table";
-import { getUserData,getUserDataById, postUser, deleteUser } from "../../actions/user_actions";
+import { getUserData, getUserDataById, postUser, deleteUser } from "../../actions/user_actions";
 import ModalAddUpdate from './modalAddUpdate';
 class ManageUser extends Component {
     state = {
@@ -12,15 +12,15 @@ class ManageUser extends Component {
         searchQuery: "",
         modalShow: false,
         modalUser: ""
-      };
-      columns = [
+    };
+    columns = [
         {
-          path: "fullName",
-          label: "FullName"
+            path: "fullName",
+            label: "FullName"
         },
         {
-          path: "email",
-          label: "Email"
+            path: "email",
+            label: "Email"
         },
         {
             path: "phoneNumber",
@@ -28,40 +28,43 @@ class ManageUser extends Component {
         },
         {
             path: "status",
-            label: "Status"  
+            label: "Status"
         },
         {
             key: "update",
             content: fac => (
                 <React.Fragment>
-                <button
-                    onClick={() => this.showUpdateForm(fac)}
-                    className="btn btn-primary btn-sm"
-                >
-                    <i className="fa fa-pen" />
-                </button>
-                <button
-                    onClick={() => this.handleDelete(fac)}
-                    className="btn btn-danger btn-sm"
-                >
-                    <i className="fa fa-trash" />
-                </button>
-                <button
-                    onClick={() => this.handleInfor(fac)}
-                    className="btn btn-info btn-sm">
-                    <i className="fa fa-info" />
-                </button>
+                    <button
+                        onClick={() => this.showUpdateForm(fac)}
+                        className="btn btn-primary btn-sm"
+                    >
+                        <i className="fa fa-pen" />
+                    </button>
+                    <button
+                        onClick={() => this.handleDelete(fac)}
+                        className="btn btn-danger btn-sm"
+                    >
+                        <i className="fa fa-trash" />
+                    </button>
+                    <button
+                        onClick={() => this.handleInfor(fac)}
+                        className="btn btn-info btn-sm">
+                        <i className="fa fa-info" />
+                    </button>
                 </React.Fragment>
             )
         }
     ];
     handleDelete = user => {
-        this.props.dispatch(deleteUser(user.id)).then(res=>this.props.dispatch(getUserData()));
+        var result = window.confirm("Do you want delete this user?");
+        if(result){
+            this.props.dispatch(deleteUser(user.id)).then(res => this.props.dispatch(getUserData()));
+        }  
     };
-    handleSubmit = (userSubmit,userId) => {
-        this.props.dispatch(postUser(userSubmit,userId)).then(res=>this.props.dispatch(getUserData()));
+    handleSubmit = (userSubmit, userId) => {
+        this.props.dispatch(postUser(userSubmit, userId)).then(res => this.props.dispatch(getUserData()));
     }
-    async showUpdateForm(user){
+    async showUpdateForm(user) {
         await this.props.dispatch(getUserDataById(user.id));
         console.log(this.props.user)
         this.setState({
@@ -70,76 +73,98 @@ class ManageUser extends Component {
         });
     }
 
-    async handleInfor(user){
+    async handleInfor(user) {
         await this.props.dispatch(getUserDataById(user.id));
         this.setState({
-         modalShow: !this.state.modalShow,
-         modalUser: this.props.user,
+            modalShow: !this.state.modalShow,
+            modalUser: this.props.user,
         });
-     }
+    }
     componentDidMount = async () => {
         await this.props.dispatch(getUserData());
-        this.setState({user: this.props.users})
+        this.setState({ user: this.props.users })
     }
     toggle = () => {
         this.setState({
             modalShow: !this.state.modalShow,
             modalUser: ""
         });
-        
+
     };
     getData = () => {
-        const {pageSize, currentPage, searchQuery} = this.state;
+        const { pageSize, currentPage, searchQuery } = this.state;
+        let filtered = this.props.users.data;
+        if (searchQuery) {
+            filtered = this.props.users.data.filter(m =>
+                m.fullName.toLowerCase().startsWith(searchQuery.toLowerCase())
+            );
+        }
         const dataPagination = paginate(
-            this.props.users.data,
+            filtered,
             currentPage,
             pageSize
         )
-        return {data: dataPagination};
+        return { data: dataPagination, itemsCount: filtered.length };
     };
 
-    handlePageChange = (page) =>{
+    handlePageChange = (page) => {
         this.setState({
             currentPage: page
         })
-    }
+    };
 
+    handleSearch = event => {
+        const element = event.target.value;
+        this.setState({
+            searchQuery: element
+        });
+    };
     render() {
         const {
             pageSize,
             currentPage,
-            searchQuery,
             modalUser,
             modalShow,
-          } = this.state;
-          const itemsCount = this.props.users.data.length;
-          const{data: dataPagination} = this.getData();
+        } = this.state;
+        const { data: dataPagination, itemsCount: itemsCount } = this.getData();
         return (
             <div style={{ marginLeft: "50px" }}>
-                <h1 className="h3 mb-2 text-gray-800">Manage User</h1>
-                <button
-                onClick = {this.toggle}
-                >Add new User
-                </button>
+                <h1 className="h3 mb-2 text-gray-800 text-center">Manage User</h1>
+                <div className="row justify-content-end">
+                    <div className="col-6">
+                        <input
+                            placeholder="Search User Name"
+                            type="text"
+                            className="form-control"
+                            onChange={this.handleSearch}
+                        /></div>
+                    <div className="col-4">
+                        <button
+                            onClick={this.toggle}
+                            className="btn btn-info">
+                            Add new User
+                         </button>
+                    </div>
+                </div>
                 <ModalAddUpdate
-                show={modalShow}
-                toggle={this.toggle}
-                userInfo={modalUser}
-                onSubmit={this.handleSubmit}
+                    show={modalShow}
+                    toggle={this.toggle}
+                    userInfo={modalUser}
+                    onSubmit={this.handleSubmit}
                 />
-                <Table data={dataPagination} columns={this.columns}/>
+                <Table data={dataPagination} columns={this.columns} />
                 <Pagination
-                itemsCount={itemsCount}
-                pageSize={pageSize}
-                currentPage={currentPage}
-                onPageChange={this.handlePageChange}
+                    itemsCount={itemsCount}
+                    pageSize={pageSize}
+                    currentPage={currentPage}
+                    onPageChange={this.handlePageChange}
                 />
             </div>
         );
     }
 }
 const mapStateToProps = state => {
-    return { users: state.user, user: state.user.dataById};
+    return { users: state.user, user: state.user.dataById };
 };
 
 export default connect(mapStateToProps)(ManageUser);
